Refactor conda_store_server.conda -> conda_store_server.conda_utils
