+++ conda-store-server/conda_store_server/action/get_conda_prefix_stats.py
@@ -1,12 +1,12 @@
 import pathlib
 
-from conda_store_server import action, conda, utils
+from conda_store_server import action, conda_utils, utils
 
 
 @action.action
 def action_get_conda_prefix_stats(context, conda_prefix: pathlib.Path):
     # safeguard to try and prevent destructive actions
-    if not conda.is_conda_prefix(conda_prefix):
+    if not conda_utils.is_conda_prefix(conda_prefix):
         raise ValueError("given prefix is not a conda environment")
 
     stats = {}
