+++ conda-store-server/conda_store_server/action/remove_conda_prefix.py
@@ -1,7 +1,7 @@
 import shutil
 import pathlib
 
-from conda_store_server import conda, action
+from conda_store_server import conda_utils, action
 
 
 @action.action
@@ -10,7 +10,7 @@ def action_remove_conda_prefix(
     conda_prefix: pathlib.Path,
 ):
     # safeguard to try and prevent destructive actions
-    if not conda.is_conda_prefix(conda_prefix):
+    if not conda_utils.is_conda_prefix(conda_prefix):
         raise ValueError("given prefix is not a conda environment")
 
     context.log.info(f'removing conda prefix "{conda_prefix.resolve()}"')
