+++ conda-store-server/conda_store_server/schema.py
@@ -7,7 +7,7 @@ import functools
 
 from pydantic import BaseModel, Field, constr, validator
 
-from conda_store_server.conda import conda_platform
+from conda_store_server import conda_utils
 
 
 def _datetime_factory(offset: datetime.timedelta):
@@ -210,7 +210,7 @@ class Settings(BaseModel):
     )
 
     conda_platforms: List[str] = Field(
-        ["noarch", conda_platform()],
+        ["noarch", conda_utils.conda_platform()],
         description="Conda platforms to download package repodata.json from. By default includes current architecture and noarch",
         metadata={"global": True},
     )
@@ -241,7 +241,7 @@ class Settings(BaseModel):
     )
 
     conda_solve_platforms: List[str] = Field(
-        [conda_platform()],
+        [conda_utils.conda_platform()],
         description="Conda platforms to solve environments for via conda-lock. Must include current platform.",
         metadata={"global": False},
     )
